fixed Component

+++ src/main/java/dev/nerobeev/economic_cybernetics_system/dto/component/ComponentResponse.java
@@ -6,7 +6,6 @@ import dev.nerobeev.economic_cybernetics_system.domain.UnitOfMeasure;
 import dev.nerobeev.economic_cybernetics_system.dto.material.MaterialResponse;
 import dev.nerobeev.economic_cybernetics_system.dto.product.ProductResponse;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
@@ -15,10 +14,10 @@ public record ComponentResponse(
         String name,
         String uCode,
         UnitOfMeasure unit,
-        BigDecimal costPerUnit,
-        BigDecimal pricePerUnit,
+        Long costPerUnit,
+        Long pricePerUnit,
         String producer,
-        BigDecimal quantity,
+        Long quantity,
         Status status,
         IndustryCode industryCode,
         String planPeriod,
