+++ src/main/java/dev/nerobeev/economic_cybernetics_system/dto/material/MaterialCreateRequest.java
@@ -4,16 +4,15 @@ import dev.nerobeev.economic_cybernetics_system.domain.IndustryCode;
 import dev.nerobeev.economic_cybernetics_system.domain.Status;
 import dev.nerobeev.economic_cybernetics_system.domain.UnitOfMeasure;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record MaterialCreateRequest(
     String name,
     UnitOfMeasure unit,
-    BigDecimal costPerUnit,
-    BigDecimal pricePerUnit,
+    Long costPerUnit,
+    Long pricePerUnit,
     String producer,
-    BigDecimal quantity,
+    Long quantity,
     Status status,
     IndustryCode industryCode,
     String planPeriod,
