+++ src/main/java/dev/nerobeev/economic_cybernetics_system/dto/material/MaterialResponse.java
@@ -4,7 +4,6 @@ import dev.nerobeev.economic_cybernetics_system.domain.IndustryCode;
 import dev.nerobeev.economic_cybernetics_system.domain.Status;
 import dev.nerobeev.economic_cybernetics_system.domain.UnitOfMeasure;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record MaterialResponse(
@@ -12,10 +11,10 @@ public record MaterialResponse(
     String name,
     String uCode,
     UnitOfMeasure unit,
-    BigDecimal costPerUnit,
-    BigDecimal pricePerUnit,
+    Long costPerUnit,
+    Long pricePerUnit,
     String producer,
-    BigDecimal quantity,
+    Long quantity,
     Status status,
     IndustryCode industryCode,
     String planPeriod,
