+++ src/main/java/dev/nerobeev/economic_cybernetics_system/dto/product/ProductCreateRequest.java
@@ -3,7 +3,6 @@ package dev.nerobeev.economic_cybernetics_system.dto.product;
 import dev.nerobeev.economic_cybernetics_system.domain.IndustryCode;
 import dev.nerobeev.economic_cybernetics_system.domain.Status;
 import dev.nerobeev.economic_cybernetics_system.domain.UnitOfMeasure;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 // Request для создания продукта
@@ -11,10 +10,10 @@ import java.time.LocalDate;
 public record ProductCreateRequest(
         String name,
         UnitOfMeasure unit,
-        BigDecimal costPerUnit,
-        BigDecimal pricePerUnit,
+        Long costPerUnit,
+        Long pricePerUnit,
         String producer,
-        BigDecimal quantity,
+        Long quantity,
         Status status,
         IndustryCode industryCode,
         String planPeriod,
