+++ src/main/java/dev/nerobeev/economic_cybernetics_system/dto/production/ProductionCostCreateRequest.java
@@ -1,12 +1,10 @@
 package dev.nerobeev.economic_cybernetics_system.dto.production;
 
-import java.math.BigDecimal;
-
 public record ProductionCostCreateRequest(
     String name,
-    BigDecimal energyCost,
-    BigDecimal laborHours,
-    BigDecimal equipmentCost
+    Long energyCost,
+    Long laborHours,
+    Long equipmentCost
 
 ) {
 }
