+++ src/main/java/dev/nerobeev/economic_cybernetics_system/dto/production/ProductionCostResponse.java
@@ -1,13 +1,12 @@
 package dev.nerobeev.economic_cybernetics_system.dto.production;
 
-import java.math.BigDecimal;
 
 public record ProductionCostResponse(
     Long id,
     String name,
-    BigDecimal energyCost,
-    BigDecimal laborHours,
-    BigDecimal equipmentCost
+    Long energyCost,
+    Long laborHours,
+    Long equipmentCost
 
 ) {
 }
