+++ src/main/java/dev/nerobeev/economic_cybernetics_system/entity/ProductionCost.java
@@ -4,8 +4,6 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Entity
 @Table(name = "production_costs")
 @NoArgsConstructor
@@ -24,63 +22,63 @@ public class ProductionCost {
 
   @NotNull
   @Column(name = "energy_cost")
-  private BigDecimal energyCost;           // Стоимость энергии
+  private Long energyCost;           // Стоимость энергии
 
   @NotNull
   @Column(name = "labor_hours")
-  private BigDecimal laborHours;           // Зарплата
+  private Long laborHours;           // Зарплата
 
   @NotNull
   @Column(name = "equipment_cost")
-  private BigDecimal equipmentCost;        // Стоимость оборудования
+  private Long equipmentCost;        // Стоимость оборудования
 
   @NotNull
   @Column(name = "material_cost")
-  private BigDecimal materialCost;         // Стоимость расходных материалов
+  private Long materialCost;         // Стоимость расходных материалов
 
   @Column(name = "logistics_cost")
-  private BigDecimal logisticsCost;        // Стоимость доставки
+  private Long logisticsCost;        // Стоимость доставки
 
   @NotNull
   @Column(name = "license_cost")
-  private BigDecimal licenseCost;          // Стоимость лицензий
+  private Long licenseCost;          // Стоимость лицензий
 
   @NotNull
   @Column(name = "tax_cost")
-  private BigDecimal taxCost;              // Налоги
+  private Long taxCost;              // Налоги
 
   @NotNull
   @Column(name = "social_cost")
-  private BigDecimal socialCost;           // Социальные выплаты
+  private Long socialCost;           // Социальные выплаты
 
   @NotNull
   @Column(name = "amortization_cost")
-  private BigDecimal amortizationCost;     // Амортизация
+  private Long amortizationCost;     // Амортизация
 
   @NotNull
   @Column(name = "equipment_maintenance_cost")
-  private BigDecimal equipmentMaintenanceCost; // Обслуживание оборудования
+  private Long equipmentMaintenanceCost; // Обслуживание оборудования
 
   @NotNull
   @Column(name = "administrative_cost")
-  private BigDecimal administrativeCost;   // Административные расходы
+  private Long administrativeCost;   // Административные расходы
 
   @Column(name = "rental_cost")
-  private BigDecimal rentalCost;           // Аренда
+  private Long rentalCost;           // Аренда
 
   @Column(name = "communication_cost")
-  private BigDecimal communicationCost;    // Связь
+  private Long communicationCost;    // Связь
 
   @Column(name = "insurance_cost")
-  private BigDecimal insuranceCost;        // Страхование
+  private Long insuranceCost;        // Страхование
 
   @Column(name = "research_and_development_cost")
-  private BigDecimal researchAndDevelopmentCost; // НИОКР
+  private Long researchAndDevelopmentCost; // НИОКР
 
   @Column(name = "interest_cost")
-  private BigDecimal interestCost;         // Кредиты и займы, расходы на маркетинг
+  private Long interestCost;         // Кредиты и займы, расходы на маркетинг
 
   @Column(name = "eco_cost")
-  private BigDecimal ecoCost;              // Экологический сбор
+  private Long ecoCost;              // Экологический сбор
 
 }
