+++ src/main/java/dev/nerobeev/economic_cybernetics_system/repository/ProductionCostRepository.java
@@ -4,11 +4,10 @@ import dev.nerobeev.economic_cybernetics_system.entity.ProductionCost;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
 public interface ProductionCostRepository extends JpaRepository<ProductionCost,Long> {
   Optional<ProductionCost>findById(Long id);
-  Optional<ProductionCost>findByEnergyCost(BigDecimal energyCost);
+  Optional<ProductionCost>findByEnergyCost(Long energyCost);
 }
